requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
python-dotenv==1.0.0
boto3==1.34.0
gunicorn==21.2.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
